requirements.txt
typer
label-studio-sdk==1.0.3
pyyaml
yandexcloud
mergedeep